Out-of-source build.
This should help when tinkering. Deleting one directory with build
output is much easier then hunting the build output all around. This
could also help with build of additional configurations, e.g. due to
modified build flags.

So far, go with `%{_vpath_builddir}` which is generally used for this
purpose. If there is need for more build configuration, the macro can be
modified or replaces. But for now, YAGNI.